Create solution_2
We have constraints saying "without using any built-in hash table libraries" and "At most 10^4 calls will be made to put, get, and remove". Given that we might have the same hash key, we can't use Set , so seems like simply we should use Array with Linked List. I used linked list because basically Array can keep one data in a place(index). But there is possibility that we have the same hash key with different data. In that case, we need to put multiple data in the same place. That's why we need to connect each data with links.